poc-application-code/bigquery-functions/bq-table-stream-update/requirements.txt
google-cloud-pubsub
google-cloud-storage
futures